Refactoring

+++ omap/utils/binary.h
@@ -1,6 +1,7 @@
 #ifndef BINARY_H
 #define BINARY_H
 
+#include <stdint.h>
 
 #define binary(x) S_to_binary_(#x)
 
@@ -11,9 +12,9 @@ typedef volatile unsigned int byte;
 	turns a string representation of a binary number into TODO change return type?
 
 */
-static inline unsigned long long S_to_binary_(const char *s)
+static inline uint64_t S_to_binary_(const char *s)
 {
-        unsigned long long i = 0;
+        uint64_t i = 0;
         while (*s) {
                 i <<= 1;
                 i += *s++ - '0';
